Updated the JuncSalvagerSplitMatrix to correct an error from not finding the intermediary files.

+++ src/rnaseq/splicing/juncsalvager/JuncSalvagerSplitMatrixCandidates.java
@@ -134,18 +134,25 @@ public class JuncSalvagerSplitMatrixCandidates {
 				String[] arguments2 = {(outputFolderPCGP + "/" + disease + ".txt"), "1.0", "0.1", (outputFolderPCGP + "/" + disease + "_1FPKM.txt")};
 				FilterMatrixExpression.execute(arguments2);
 				rank_norm(outputFolderPCGP + "/" + disease + "_1FPKM.txt", outputFolderPCGP + "/" + disease + "_rank_1FPKM_tmp.txt", outputFolderPCGP + "/" + disease + "_rank_1FPKM_median_tmp.txt");
-				
+								 
 				String[] arguments3 = {(outputFolderPCGP + "/" + disease + "_rank_1FPKM_tmp.txt"), (outputFolderPCGP + "/" + disease + "_rank_1FPKM.txt")};
 				RemoveQuotations.execute(arguments3);
 				
+				if (!new File(outputFolderPCGP + "/" + disease + "_rank_1FPKM_tmp.txt").exists()) {
+					new File(outputFolderPCGP + "/" + disease + "_rank_1FPKM_tmp.txt").createNewFile();
+				}
 				String[] arguments4 = {outputFolderPCGP + "/" + disease + "_rank_1FPKM_median_tmp.txt", outputFolderPCGP + "/" + disease + "_rank_1FPKM_median.txt"};
 				RemoveQuotations.execute(arguments4);				
 				
 				f = new File(outputFolderPCGP + "/" + disease + "_rank_1FPKM_tmp.txt");
-				f.delete();
+				if (f.exists()) {
+					f.delete();
+				}
 				
 				f = new File(outputFolderPCGP + "/" + disease + "_rank_1FPKM_median_tmp.txt");
-				f.delete();
+				if (f.exists()) {
+					f.delete();
+				}
 				
 			}
